Add competition mutation and repository files
Added several new files for GraphQL test mutations regarding competition features, as well as respective repository interfaces for managing competition score data. Also implemented recalculation of competition ranking in service layer and refined various tests for competition functionalities.

+++ src/main/java/nl/shootingclub/clubmanager/model/CompetitionScorePoint.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class CompetitionScorePoint extends CompetitionScore<CompetitionScorePoint> {
 
     @Column(name = "score")
-    private int score;
+    private long score;
 
     public CompetitionScorePoint() {
 
@@ -22,7 +22,7 @@ public class CompetitionScorePoint extends CompetitionScore<CompetitionScorePoin
 
     @Override
     public int compareTo(CompetitionScorePoint o) {
-        return Integer.compare(this.score, o.score);
+        return Long.compare(this.score, o.score);
     }
 
     @Override
